requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Pillow==9.4.0
numpy==1.23.5
tensorflow==2.11.0
scikit-learn==1.2.1
gunicorn==20.1.0 